// govco-pwa/src/app/transversales/models/footer-models/panel-tercera-columna.ts
export interface PanelTerceraColumna {
    titulo: string;
    tituloUrlImagen: string;
    tituloTelefono: string;
    telefonoNacional: string;
    telefonoLocal: string;
    centroRelevoUrlDestino: string;
    centroRelevoTexto: string;
    lineaAnticorrupcion: string;
    correoTexto: string;
    correoInstitucional: string;
    correoUrlDestino: string;
    soLlamadaTexto: string;
    soLlamadaUrlDestino: string,
    llamadaWebTexto: string;
    llamadaWebUrlDestino: string;
    hablemosTexto: string;
    hablemosUrlDestino: string;
}